moving code from `src/raw/templatee/parse.rs` to `src/raw/template/wire.rs` (under `wire::encode_spawn_app`).
