+++ crates/svm-app/src/raw/mod.rs
@@ -10,7 +10,7 @@ mod varuint14;
 mod version;
 
 pub use app::{encode_spawn_app, parse_app};
-pub use template::{encode_deploy_template, parse_template};
+pub use template::{decode_deploy_template, encode_deploy_template};
 pub use transaction::parse_app_tx;
 
 pub use app::{DefaultAppDeserializer, DefaultAppSerializer};
